pyproject.toml
[project]
name = "agentstudy"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "llm>=0.28",
    "prompts>=0.0.1",
    "python-dotenv>=1.2.1",
    "pyviz-comms>=3.0.6",
]
